requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-socketio>=5.11,<6.0
pydantic-settings>=2.4,<3.0
redis>=5.0,<6.0
httpx>=0.27,<1.0
deep-translator>=1.11,<2.0
pytest>=8.0,<9.0
pytest-asyncio>=0.24,<1.0
